% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{baltable}
\alias{baltable}
\title{Calculate covariate balance statistics}
\usage{
baltable(Xmat, TA, wts = NULL, show_unweighted = T, n_digits = 3)
}
\arguments{
\item{Xmat}{A matrix of covariates}

\item{TA}{A vector of treatment assignments (i.e, 1 if treated, 0 in control)}

\item{wts}{A vector of estimated (or true) weights}

\item{show_unweighted}{Query if the the balance table should contain unweighted estimates}

\item{n_digits}{Number of digits to print for the table}
}
\value{
A table of covariate balance statistics
}
\description{
Calculate covariate balance statistics
}
\examples{

}
