% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dw99cps3}
\alias{dw99cps3}
\title{Lalonde Data Set - Current Population Survey Data: Second Subset}
\format{
A data frame with 313 rows and 10 variables:
\describe{
  \item{TA}{treatment assignment}
  \item{age}{Age}
  \item{education}{Level of Education}
  \item{black}{Black Indicator}
  \item{hispanic}{Hispanic Indicator}
  \item{married}{Indicator of Married}
  \item{nodegree}{Indicator of "No Degree"}
  \item{RE74}{Retrospective Earnings in 1974}
  \item{RE75}{Retrospective Earnings in 1975}
  \item{RE78}{Outcome Variable - Retrospective Earnings in 1978}
}
}
\source{
\url{https://users.nber.org/~rdehejia/data/.nswdata2.html}
}
\usage{
dw99cps3
}
\description{
From \url{https://users.nber.org/~rdehejia/data/.nswdata2.html}
"...Data are drawn from a paper by Robert Lalonde, "Evaluating the Econometric Evaluations of Training Programs," American Economic Review, Vol. 76, pp. 604-620. "
}
\keyword{datasets}
