% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpars_default_pars.R
\name{ebpars_default_binary}
\alias{ebpars_default_binary}
\title{Create list of default parameters for entbal - binary}
\usage{
ebpars_default_binary(estimand = "ATE", which_z = NULL)
}
\arguments{
\item{estimand}{default: \code{ATE}, alternative is \code{ATT}}

\item{which_z}{if \code{estimand} is \code{ATT}, then set \code{which_z} to target value from treatment vector}
}
\description{
Create list of default parameters for entbal - binary
}
