% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpars_default_pars.R
\name{ebpars_default_cont}
\alias{ebpars_default_cont}
\title{Create list of default parameters for entbal - continuous}
\usage{
ebpars_default_cont()
}
\description{
Create list of default parameters for entbal - continuous
}
