% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entbal-package.R
\docType{package}
\name{entbal-package}
\alias{entbal}
\alias{entbal-package}
\title{entbal: An Alternative Implementation of Entropy Balancing Weights}
\description{
An alternative implementation of entropy balancing weights.  See 
    Hainmueller (2012) for further details of the original method and information on the ebal package.
}
\keyword{propensity}
\keyword{score}
