% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{ESS}
\alias{ESS}
\title{Covariate Balance Function - Effective Sample Size}
\usage{
ESS(w)
}
\arguments{
\item{w}{weights}
}
\value{
Effective Sample Size from Weights.
}
\description{
Covariate Balance Function - Effective Sample Size
}
