% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{est_binary_te}
\alias{est_binary_te}
\title{Simple Treatment Effect Estimation}
\usage{
est_binary_te(YO, TI, wts)
}
\arguments{
\item{YO}{A Vector of Observed Outcomes}

\item{TI}{Indicators where 1 in group and zero otherwise. Typically treatment/exposure indicators.}

\item{wts}{Weights.}
}
\value{
Weighted Group Sample Variance
}
\description{
Simple difference in weighted expectations between the two groups defined by a vector of indicators.
}
