% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_wts}
\alias{extract_wts}
\title{Returns Object Weights}
\usage{
extract_wts(object)
}
\arguments{
\item{object}{object of class \code{entbal_multiclass} or \code{entbal_binary}}
}
\description{
Returns Object Weights
}
