% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{group_ESS}
\alias{group_ESS}
\title{Covariate Balance Function - Group Effective Sample Size}
\usage{
group_ESS(w, TI)
}
\arguments{
\item{w}{weights}

\item{TI}{Indicators where 1 in group and zero otherwise. Typically treatment/exposure indicators.}
}
\value{
Effective Sample of Group Defined by \code{TI==1}
}
\description{
Covariate Balance Function - Group Effective Sample Size
}
