% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ks_test}
\alias{ks_test}
\title{Weighted KS Statistic}
\usage{
ks_test(X, TA, wts = rep(1, length(X)), n_pts = 100)
}
\arguments{
\item{X}{covariate vector}

\item{TA}{treatment indicator}

\item{wts}{weights}

\item{n_pts}{number of points to evaluate weighted ecdf}
}
\description{
Weighted KS Statistic
}
