% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{log_sd_ratio}
\alias{log_sd_ratio}
\title{Covariate Balance Function - Ratio of Weighted Standard Deviations}
\usage{
log_sd_ratio(X, TA, wts)
}
\arguments{
\item{X}{A Vector of Observed Covariates}

\item{TA}{Indicators where 1 in group and zero otherwise. Typically treatment/exposure indicators.}

\item{wts}{Weights.}
}
\value{
Ratio of Weighted Standard Deviations
}
\description{
Covariate Balance Function - Ratio of Weighted Standard Deviations
}
