% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{make_wts_ATE}
\alias{make_wts_ATE}
\title{Given Propensity Scores and Binary Treatment Indicators, Create ATE Weights.}
\usage{
make_wts_ATE(ps, TI)
}
\arguments{
\item{ps}{Propensity Scores from Binary Treatment}

\item{TI}{Treatment Indicators where 1 if "exposed" and zero otherwise}
}
\value{
ATE Weights
}
\description{
Given Propensity Scores and Binary Treatment Indicators, Create ATE Weights.
}
