% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{make_wts_ATO}
\alias{make_wts_ATO}
\title{Given Propensity Scores and Binary Treatment Indicators, Create ATO Weights.}
\usage{
make_wts_ATO(ps, TI)
}
\arguments{
\item{ps}{Propensity Scores from Binary Treatment}

\item{TI}{Treatment Indicators where 1 if "exposed" and zero otherwise}
}
\value{
ATO Weights
}
\description{
Given Propensity Scores and Binary Treatment Indicators, Create ATO Weights.
}
