% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{make_wts_ATT}
\alias{make_wts_ATT}
\title{Given Propensity Scores and Binary Treatment Indicators, Create ATT Weights.}
\usage{
make_wts_ATT(ps, TI)
}
\arguments{
\item{ps}{Propensity Scores from Binary Treatment}

\item{TI}{Treatment Indicators where 1 if "exposed" and zero otherwise}
}
\value{
ATT Weights
}
\description{
Given Propensity Scores and Binary Treatment Indicators, Create ATT Weights.
}
