% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_Xmat}
\alias{make_Xmat}
\title{Make Design Matrix for Binary/Multiclass Entropy Balancing}
\usage{
make_Xmat(X, m = 1)
}
\arguments{
\item{X}{matrix of covariates to balance}

\item{m}{number of moments}
}
\description{
Make Design Matrix for Binary/Multiclass Entropy Balancing
}
