% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_balancing_continuous.R
\name{makeC2}
\alias{makeC2}
\title{Make Design Matrix for Continuous Entropy Balancing}
\usage{
makeC2(XD, A, n_moments = 3)
}
\arguments{
\item{XD}{matrix of covariates to balance}

\item{A}{vector of treatments}

\item{n_moments}{number of moments}
}
\description{
Make Design Matrix for Continuous Entropy Balancing
}
