% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_baltables}
\alias{print_baltables}
\title{Binary Balance Table}
\usage{
print_baltables(X, TA, wts, show_unweighted = TRUE, n_digits = 2)
}
\arguments{
\item{X}{Matrix of covariates}

\item{TA}{Binary indicator}

\item{wts}{Weights}

\item{show_unweighted}{Show unweighted balanced statistics? default \code{TRUE}}

\item{n_digits}{How many digits to round table, default 2}
}
\description{
Binary Balance Table
}
