% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{std_diff_means}
\alias{std_diff_means}
\title{Covariate Balance Function - Standardized Difference in Means}
\usage{
std_diff_means(X, TA, wts)
}
\arguments{
\item{X}{A Vector of Observed Covariates}

\item{TA}{Indicators where 1 in group and zero otherwise. Typically treatment/exposure indicators.}

\item{wts}{Weights.}
}
\value{
Standardized Difference in Means
}
\description{
Covariate Balance Function - Standardized Difference in Means
}
