% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{wcor}
\alias{wcor}
\title{Covariate Balance Function - Weighted Correlation}
\usage{
wcor(wts, x, y)
}
\arguments{
\item{wts}{weights}

\item{x}{first vector of covariates}

\item{y}{second vector of covariates}
}
\value{
weighted correlation between x and y
}
\description{
Covariate Balance Function - Weighted Correlation
}
