% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{wmean}
\alias{wmean}
\title{Covariate Balance Function - Weighted Mean}
\usage{
wmean(wts, x)
}
\arguments{
\item{wts}{weights}

\item{x}{vector of covariates}
}
\value{
weighted mean of x
}
\description{
Covariate Balance Function - Weighted Mean
}
