% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{wtd_moment}
\alias{wtd_moment}
\title{Given Propensity Scores, Group Indicators, and Weights: Estimate Group Non-Central Moment}
\usage{
wtd_moment(X, TI, w, mom = 1)
}
\arguments{
\item{X}{A Single Vector of Covariate Information}

\item{TI}{Indicators where 1 in group and zero otherwise. Typically treatment/exposure indicators.}

\item{w}{Weights.}

\item{mom}{Power of Moment, default to 1.

Useful for covariate balance where the groups are defined by exposure levels.}
}
\value{
Weighted Group Non-Central Moment
}
\description{
Given Propensity Scores, Group Indicators, and Weights: Estimate Group Non-Central Moment
}
