% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{wtd_sd2}
\alias{wtd_sd2}
\title{Given Propensity Scores, Group Indicators, and Weights: Estimate Weighted Group Sample Variance}
\usage{
wtd_sd2(X, TI, w)
}
\arguments{
\item{X}{A Single Vector of Covariate Information}

\item{TI}{Indicators where 1 in group and zero otherwise. Typically treatment/exposure indicators.}

\item{w}{Weights.

Useful for covariate balance where the groups are defined by exposure levels.}
}
\value{
Weighted Group Sample Variance
}
\description{
Given Propensity Scores, Group Indicators, and Weights: Estimate Weighted Group Sample Variance
}
