% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{wvar}
\alias{wvar}
\title{Covariate Balance Function - Weighted Variance}
\usage{
wvar(wts, x)
}
\arguments{
\item{wts}{weights}

\item{x}{vector of covariates}
}
\value{
weighted variance of x
}
\description{
Covariate Balance Function - Weighted Variance
}
