% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wtd_moment}
\alias{wtd_moment}
\title{Calculate an estimate of the weighted raw moment \eqn{E(X^{p})} conditioned upon a vector of group indicators}
\usage{
wtd_moment(X, TI, w, mom = 1L)
}
\arguments{
\item{X}{Vector of observed covariate values}

\item{TI}{Vector representing group membership, i.e., \code{TI = (0, 1, 0, ...)}, where 1 indicates group membership}

\item{w}{Vector of weights, these need not sum to one in each group and are normalized in the equation}

\item{mom}{Power \eqn{p} of the required moment estimate}
}
\value{
Estimate of the weighted raw moment in the group represented by \code{TI}
}
\description{
Calculate an estimate of the weighted raw moment \eqn{E(X^{p})} conditioned upon a vector of group indicators
}
