% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wtd_sd2}
\alias{wtd_sd2}
\title{Calculate a weighted variance conditioned upon a vector of group indicators}
\usage{
wtd_sd2(X, TI, w)
}
\arguments{
\item{X}{Vector of observed covariate values}

\item{TI}{Vector representing group membership, i.e., \code{TI = (0, 1, 0, ...)}, where 1 indicates group membership}

\item{w}{Vector of weights, these need not sum to one in each group and are normalized in the equation}
}
\value{
Estimate of the weighted variance in the group represented by \code{TI}
}
\description{
Calculate a weighted variance conditioned upon a vector of group indicators
}
